DB 변경사항 적용 & 자유게시판 수정 삭제 기능 추가
- MySQL/MariaDB 사용시 TIMESTAMP 컬럼에 default값을 미지정할 경우 on update 트리거가 자동생성됩니다. 이는 의도치 않은 날짜 변경을 야기하기에 default값을 명시함으로써 자동 갱신 트리거를 해제합니다.
- CASCADE 구문을 제거합니다.
- DB변경사항을 sql-generate.sql 파일에 반영합니다.
- 게시글 삭제 기능을 추가합니다.
- 게시글 수정시 마지막으로 추가된 파일그룹의 최소 조회권한이 SECRET(비밀글)으로 고정되는 문제를 수정합니다.

+++ src/main/java/kr/pe/hayarobys/nox/common/comment/CommentServiceImpl.java
@@ -35,6 +35,8 @@ public class CommentServiceImpl implements CommentService{
 		commentDAO.deleteCommentGroupByCommentGroupNo(commentGroupNo);
 		
 		// 파일 그룹과 그에 연결된 DB파일, 물리파일 제거 서비스 요청
-		uploadService.deleteFileGroupByFileGroupNoList(fileGroupNoList);
+		if(fileGroupNoList.size() > 0){
+			uploadService.deleteFileGroupByFileGroupNoList(fileGroupNoList);
+		}
 	}
 }
